requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
bidict==0.22.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
dnspython==2.2.1
eventlet==0.33.1
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
pytest==7.1.2
python-engineio==4.3.2
python-socketio==5.6.0
requests==2.28.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
websocket-client==1.3.2
zope.event==4.5.0
zope.interface==5.4.0
